🚚 move tab2.5 (so tab 3) to separate module
PlotDataPoint has only static methods and should be a moduel, but this is another discussion, see
issue #834
