+++ webinterface/pages/base_pages/quant.py
@@ -1,21 +1,16 @@
 """Streamlit-based web interface for ProteoBench."""
 
 import copy
-import glob
 import json
 import logging
 import os
 import tempfile
 import uuid
-import zipfile
-from datetime import datetime
 from typing import Any, Dict, Optional
 
 import pages.texts.proteobench_builder as pbb
 import pandas as pd
-import plotly.graph_objects as go
 import streamlit as st
-import streamlit_utils
 from pages.pages_variables.Quant.lfq_DDA_ion_QExactive_variables import (
     VariablesDDAQuant,
 )
@@ -29,9 +24,10 @@ from proteobench.modules.quant.quant_lfq_ion_DDA_QExactive import (
 )
 from proteobench.plotting.plot_quant import PlotDataPoint
 
-logger: logging.Logger = logging.getLogger(__name__)
-from . import tab1_results, tab2_form_upload_data
+from . import tab1_results, tab2_form_upload_data, tab3_indepth_plots
 from .inputs import generate_input_widget
+
+logger: logging.Logger = logging.getLogger(__name__)
 
 
 def compare_dictionaries(old_dict, new_dict):
@@ -412,9 +408,11 @@ class QuantUIObjects:
         downloads_df = df[["id", "intermediate_hash"]]
         downloads_df.set_index("intermediate_hash", drop=False, inplace=True)
 
-        if self.variables_quant.placeholder_dataset_selection_container not in st.session_state.keys():
-            st.session_state[self.variables_quant.placeholder_dataset_selection_container] = st.empty()
-            st.session_state[self.variables_quant.dataset_selector_id_uuid] = uuid.uuid4()
+        key_in_state = self.variables_quant.placeholder_dataset_selection_container
+        if key_in_state not in st.session_state.keys():
+            st.session_state[key_in_state] = st.empty()
+            key_in_state = self.variables_quant.dataset_selector_id_uuid
+            st.session_state[key_in_state] = uuid.uuid4()
 
         st.subheader("Select dataset to plot")
 
@@ -431,7 +429,14 @@ class QuantUIObjects:
         )
 
         public_id, selected_hash = dataset_selection
-        self.generate_indepth_plots(True, public_id=public_id, public_hash=selected_hash)
+        tab3_indepth_plots.generate_indepth_plots(
+            variables_quant=self.variables_quant,
+            parsesettingsbuilder=self.parsesettingsbuilder,
+            user_input=self.user_input,
+            recalculate=True,
+            public_id=public_id,
+            public_hash=selected_hash,
+        )
 
     def generate_submission_button(self) -> Optional[str]:
         """
@@ -717,25 +722,6 @@ class QuantUIObjects:
                         self.variables_quant, self.user_input["input_format"], value, key, editable=editable
                     )
 
-    def generate_sample_name(self) -> str:
-        """
-        Generate a unique sample name based on the input format, software version, and the current timestamp.
-
-        Returns
-        -------
-        str
-            The generated sample name.
-        """
-        time_stamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        sample_name = "-".join(
-            [
-                self.user_input["input_format"],
-                time_stamp,
-            ]
-        )
-
-        return sample_name
-
     def get_form_values(self) -> Dict[str, Any]:
         """
         Retrieve all user inputs from Streamlit session state and returns them as a dictionary.
@@ -793,145 +779,6 @@ class QuantUIObjects:
             and pr_url is not None
         ):
             self.show_submission_success_message(pr_url)
-
-    def generate_indepth_plots(
-        self, recalculate: bool, public_id: Optional[str], public_hash: Optional[str]
-    ) -> go.Figure:
-        """
-        Generate and return plots based on the current benchmark data in Tab 2.5.
-
-        Parameters
-        ----------
-        recalculate : bool
-            Whether to recalculate the plots.
-        public_id : Optional[str], optional
-            The dataset to plot, either "Uploaded dataset" or name of public run.
-        public_hash : Optional[str], optional
-            The hash of the selected public dataset. If None, the uploaded dataset is displayed.
-
-        Returns
-        -------
-        go.Figure
-            The generated plots for the selected dataset.
-        """
-        # no uploaded dataset and no public dataset selected? nothing to plot!
-        if self.variables_quant.result_perf not in st.session_state.keys():
-            if public_hash is None:
-                st.error(":x: Please submit a result file or select a public run for display", icon="🚨")
-                return False
-            elif public_id == "Uploaded dataset":
-                st.error(":x: Please submit a result file in the Submit New Data Tab", icon="🚨")
-                return False
-
-        if public_id == "Uploaded dataset":
-            performance_data = st.session_state[self.variables_quant.result_perf]
-        else:
-            # Downloading the public performance data
-            performance_data = None
-            if st.secrets["storage"]["dir"] != None:
-                dataset_path = os.path.join(st.secrets["storage"]["dir"], public_hash)
-                # Define the path and the pattern
-                pattern = os.path.join(dataset_path, "*_data.zip")
-
-                # Use glob to find files matching the pattern
-                zip_files = glob.glob(pattern)
-
-                # Check that at least one match was found
-                if not zip_files:
-                    st.error(":x: Could not find the files on the server", icon="🚨")
-                    return
-
-                # (Optional) handle multiple matches if necessary
-                zip_path = zip_files[0]  # Assumes first match is the desired one
-
-                # Open the ZIP file and extract the desired CSV
-                with zipfile.ZipFile(zip_path) as z:
-                    with z.open("result_performance.csv") as f:
-                        performance_data = pd.read_csv(f)
-
-        # Filter the data based on the slider condition (as before)
-        performance_data = performance_data[
-            performance_data["nr_observed"] >= st.session_state[st.session_state[self.variables_quant.slider_id_uuid]]
-        ]
-
-        if recalculate:
-            parse_settings = self.parsesettingsbuilder.build_parser(self.user_input["input_format"])
-
-            fig_logfc = PlotDataPoint.plot_fold_change_histogram(
-                performance_data, parse_settings.species_expected_ratio()
-            )
-            fig_CV = PlotDataPoint.plot_CV_violinplot(performance_data)
-            fig_MA = PlotDataPoint.plot_ma_plot(
-                performance_data,
-                parse_settings.species_expected_ratio(),
-            )
-            st.session_state[self.variables_quant.fig_cv] = fig_CV
-            st.session_state[self.variables_quant.fig_logfc] = fig_logfc
-        else:
-            fig_logfc = st.session_state[self.variables_quant.fig_logfc]
-            fig_CV = st.session_state[self.variables_quant.fig_cv]
-            fig_MA = st.session_state[self.variables_quant.fig_ma_plot]
-
-        if self.variables_quant.first_new_plot:
-            col1, col2 = st.columns(2)
-            col1.subheader("Log2 Fold Change distributions by species.")
-            col1.markdown(
-                """
-                    log2 fold changes calculated from {}
-                """.format(
-                    public_id
-                )
-            )
-            col1.plotly_chart(fig_logfc, use_container_width=True)
-
-            col2.subheader("Coefficient of variation distribution in Condition A and B.")
-            col2.markdown(
-                """
-                    CVs calculated from {}
-                """.format(
-                    public_id
-                )
-            )
-            col2.plotly_chart(fig_CV, use_container_width=True)
-
-            col1.markdown("---")  # optional horizontal separator
-
-            col1.subheader("MA plot")
-            col1.markdown(
-                """
-                    MA plot calculated from {}
-                """.format(
-                    public_id
-                )
-            )
-            # Example: plot another figure or add any other Streamlit element
-            # st.plotly_chart(fig_additional, use_container_width=True)
-            col1.plotly_chart(fig_MA, use_container_width=True)
-
-        else:
-            pass
-
-        st.subheader("Sample of the processed file for {}".format(public_id))
-        st.markdown(open(self.variables_quant.description_table_md, "r").read())
-        st.session_state[self.variables_quant.df_head] = st.dataframe(performance_data.head(100))
-
-        st.subheader("Download table")
-        random_uuid = uuid.uuid4()
-        if public_id == "Uploaded dataset":
-            # user uploaded data does not have sample name yet
-            sample_name = self.generate_sample_name()
-        else:
-            # use public run name as sample name
-            sample_name = public_id
-        st.download_button(
-            label="Download",
-            data=streamlit_utils.save_dataframe(performance_data),
-            file_name=f"{sample_name}.csv",
-            mime="text/csv",
-            key=f"{random_uuid}",
-        )
-
-        return fig_logfc
 
     @st.fragment
     def display_all_data_results_main(self) -> None:
